requirements.txt
flet
escpos
python-escpos[all]
pyserial
bcrypt
python-decouple
reportlab